factory_architect/requirements.txt
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
ezdxf==1.1.4
loguru==0.7.2
python-dotenv==1.0.0
google-generativeai>=0.7.2
